Moved the streaming API to stealTools.streams
Makes it a bit nicer to use. Updated the documentation as well.
