+++ doc/concat-stream.md
@@ -1,9 +1,9 @@
-@function steal-tools.createConcatStream createConcatStream
+@function steal-tools.streams.concat concat
 @parent steal-tools.stream
 
 Create a [stream](https://nodejs.org/api/stream.html) of [steal-tools.BuildResult] objects.
 
-@signature `stealTools.createConcatStream()`
+@signature `stealTools.streams.concat()`
 
 @return {Stream<steal-tools.BuildResult>} A stream of objects that contains the dependency graph, bundles, and loader used to perform the trace.
 
@@ -13,24 +13,24 @@ Create a [stream](https://nodejs.org/api/stream.html) of [steal-tools.BuildResul
 
 The concat stream is used to **concatenate** the source code from a bundle and (optionally) create source maps.
 
-This API must be used in conjuction with [steal-tools.createMultiBuildStream].
+This API must be used in conjuction with [steal-tools.streams.transpileAndBundle].
 
 ```
-var graphStream = stealTools.createGraphStream({
+var s = require("steal-tools").streams;
+
+var graphStream = s.graph({
 	config: __dirname + "/package.json!npm"
 }, {
 	minify: false
 });
 
 var buildStream = graphStream.pipe(
-	stealTools.createMultiBuildStream()
+	s.transpileAndBundle()
 );
 
-var concatStream = buildStream.pipe(
-	stealTools.createConcatStream()
-);
+var concatStream = buildStream.pipe(s.concat());
 
-concat.on("data", function(data){
+concatStream.on("data", function(data){
 	var bundle = data.bundles[0];
 	assert(bundle.source, "The source was concatenated");
 });
