+++ doc/guides/streams.md
@@ -9,17 +9,17 @@ The streaming APIs are more powerful, but more complex, if you just want to buil
 
 The following are the streams that can be created. Each stream is explained individually and at the end of the docs we'll talking about putting them all together. Many of the examples use the [through2](https://www.npmjs.com/package/through2) module, which provide an easy way to work with Node streams.
 
-## createGraphStream
+## graph
 
 The basis for StealTool's build process is a **dependency graph**. A dependency graph is an object that contains all of your app's dependencies. The key is the name of the module and the value is a [steal-tools.node] object.
 
-**createGraphStream** is the first stream created and is required for use with all of the other streams.
+The graph stream is the first stream created and is required for use with all of the other streams.
 
 ```
-var stealTools = require("steal-tools");
+var s = require("steal-tools").streams;
 var through = require("through2");
 
-var graphStream = stealTools.createGraphStream({
+var graphStream = s.graph({
 	config: __dirname + "/package.json!npm"
 });
 
@@ -28,23 +28,20 @@ graphStream.pipe(through(function(data){
 }));
 ```
 
-## createMultiBuildStream
+## transpileAndBundle
 
-Given a graph stream, **createMultiBuildStream** does data transformation on your dependency graph. This includes **transpiling** to AMD, **minifying** (if enabled in your options), and dividing into optimized **bundles**.
+Given a graph stream, **transpileAndBundle** does data transformation on your dependency graph. This includes **transpiling** to AMD, **minifying** (if enabled in your options), and dividing into optimized **bundles**.
 
 ```
-var stealTools = require("steal-tools");
+var s = require("steal-tools").streams;
 var through = require("through2");
-
-var createGraphStream = stealTools.createGraphStream;
-var multiBuild = stealTools.createMultiBuildSteam;
 
 var system = {
 	config: __dirname + "/package.json!npm"
 };
 
-var buildStream = createGraphStream(system)
-	.pipe(multiBuild());
+var buildStream = s.graph(system)
+	.pipe(s.transpileAndBundle());
 
 buildStream.pipe(through(function(data){
 	// data contains everything from the graphStream
@@ -52,49 +49,39 @@ buildStream.pipe(through(function(data){
 }));
 ```
 
-## createConcatStream
+## concat
 
 Once you've created the multi build stream that contains all of your application's bundles, you can pipe it into a concat stream. The concat stream will **concatenate** the source from all of the graph's dependencies into a single source code.
 
 The [steal-tools.BuildResult] object is what is returned from this stream, and each of the **bundles** within will now contain a `source` property.
 
-## createWriteStream
+## write
 
-After concatentating the bundles you call **createWriteStream** to write the source to the filesystem. If you where only interested in examining the bundles in-memory you could skip this step.
+After concatentating the bundles you call **write** to write the source to the filesystem. If you where only interested in examining the bundles in-memory you could skip this step.
 
-Otherwise, pipe in the result from **createConatStream** here.
+Otherwise, pipe in the result from **concat** here.
 
 # Putting it all together
 
 Given these APIs you can construct your own build by piping the streams into each other. Here's what a typical build looks like:
 
 ```
-var stealTools = require("steal-tools");
-
-var createGraphStream = stealTools.createGraphStream;
-var multiBuild = stealTools.createMultiBuildStream;
-var concat = stealTools.createConcatStream;
-var write = stealTools.createWriteStream;
+var s = require("steal-tools").streams;
 
 var system = {
 	config: __dirname + "/package.json!npm"
 };
 
-var stream = createGraphStream(system)
-	.pipe(multiBuild())
-	.pipe(concat())
-	.pipe(write());
+var stream = s.graph(system)
+	.pipe(s.transpileAndBundle())
+	.pipe(s.concat())
+	.pipe(s.write());
 ```
 
 Now let's say we wanted to rerun [babel](https://babeljs.io/) to transpile non-es6 modules. We could do this by injecting a stream after *multiBuild* has completed:
 
 ```
-var stealTools = require("steal-tools");
-
-var createGraphStream = stealTools.createGraphStream;
-var multiBuild = stealTools.createMultiBuildStream;
-var concat = stealTools.createConcatStream;
-var write = stealTools.createWriteStream;
+var s = require("steal-tools").streams;
 
 var babel = require("babel-core");
 var transform =	function(source){
@@ -107,7 +94,7 @@ var transform =	function(source){
 	});
 };
 
-var createBabelStream = function(){
+var transform = function(){
 	return through.obj(function(data){
 		data.bundles.forEach(function(bundle){
 			bundle.nodes.forEach(function(node){
@@ -121,10 +108,10 @@ var system = {
 	config: __dirname + "/package.json!npm"
 };
 
-var stream = createGraphStream(system)
-	.pipe(multiBuild())
-	.pipe(createBabelStream())
-	.pipe(concat())
-	.pipe(write());
+var stream = s.graph(system)
+	.pipe(s.transpileAndBundle())
+	.pipe(transform())
+	.pipe(s.concat())
+	.pipe(s.write());
 ```
