+++ lib/assign_default_options.js
@@ -5,7 +5,7 @@ module.exports = function(config, options){
 	if(options.__defaultsAssigned) return options;
 
 	options = assign({ // Defaults
-		minify: true,
+		minify: false,
 		bundleSteal: false,
 		uglifyOptions: {},
 		cleanCSSOptions: {},
