+++ test/multibuild_test.js
@@ -1682,29 +1682,4 @@ describe("multi build", function(){
 			});
 		});
 	});
-
-	describe("createMultiBuildStream", function(){
-		var through = require("through2");
-
-		it("Creates a stream of BuildResult", function(done){
-			var system = {
-				config: __dirname + "/bundle/stealconfig.js",
-				main: "bundle"
-			};
-			var options = { quiet: true };
-
-			var graphStream = stealTools.createGraphStream(system, options);
-			var buildStream = graphStream.pipe(
-				stealTools.createMultiBuildStream()
-			);
-
-			buildStream.pipe(through.obj(function(data){
-				assert(data.graph, "has the graph");
-				assert(data.bundles, "has the bundles");
-				assert(data.configuration, "has the configuration");
-
-				done();
-			}));
-		});
-	});
 });
